🐍Python/fastapi-with-postgresql-alembic/utils/utils_helper.py
from typing import Optional
import jwt 
from datetime import datetime, timedelta
from config.database import engine
from model.user_model import UserDB
import os

SECRET_KEY = os.getenv("SECRET_KEY")
ALGORITHM = os.getenv("ALGORITHM")
ACCESS_TOKEN_EXPIRE_MINUTES = os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES")


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    try:
        to_encode = data.copy()
        expire = datetime.utcnow() + (expires_delta or timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES))
        to_encode.update({"exp": expire})
        return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    except Exception as e:
        return {
            "Error": str(e),
            "message": "Error creating access token",
            "Status": 500
        }

def decode_token(token: str):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload
    except Exception as e:
        return {
            "Error": str(e),
            "message": "Error decoding token",
            "Status": 500   
        }

