🐍Python/fastapi-with-postgresql-alembic/pyproject.toml
[project]
name = "fastapi-with-postgresql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "fastapi[all]>=0.115.12",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.5",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
]
